+++ src/channel.rs
@@ -109,6 +109,13 @@ impl Video {
             .map(Video::from)
             .collect()
     }
+
+    pub fn needs_update(&self, other: &Video) -> bool {
+        self.title != other.title
+            || self.length.is_none() && other.length.is_some()
+            || matches!(self.length, Some(length) if length == 0)
+                && matches!(other.length, Some(length) if length != 0)
+    }
 }
 
 impl From<&Value> for Video {